requirements.txt
django==2.2.6
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
django-cors-headers==3.1.1
requests==2.22.0